+++ gems/aws-sdk-s3/spec/object_summary_spec.rb
@@ -53,10 +53,10 @@ module Aws
         object_summary.upload_file('source', opt_name: 'opt-value')
       end
 
-      it 'delegates #upload_chunks to #object' do
+      it 'delegates #upload_stream to #object' do
         block = proc { :foo }
-        expect(object).to receive(:upload_chunks).with(opt_name: 'opt-value', &block)
-        object_summary.upload_chunks(opt_name: 'opt-value', &block)
+        expect(object).to receive(:upload_stream).with(opt_name: 'opt-value', &block)
+        object_summary.upload_stream(opt_name: 'opt-value', &block)
       end
 
       it 'delegates #download_file to #object' do